Move existing tests for Name into new pacakage
